Components: Extract a reusable post title component
