--- editor/post-title/style.scss
@@ -1,73 +1,0 @@
-.editor-visual-editor > .editor-post-title:not( [ data-align="wide" ] ),
-.editor-post-title {
-	margin-left: auto;
-	margin-right: auto;
-	position: relative;
-	max-width: $visual-editor-max-width + ( 2 * $block-mover-padding-visible );
-	padding: 5px 0;
-
-	@include break-small() {
-		padding: 5px ( $block-padding + $block-mover-padding-visible );
-	}
-
-	h1 {
-		outline: 1px solid transparent;
-		font-size: $editor-font-size;
-		transition: 0.2s outline;
-		margin-top: 0;
-		margin-bottom: 0;
-		padding: $block-padding;
-
-		@include break-small() {
-			margin-left: -$block-padding;
-			margin-right: -$block-padding;
-		}
-	}
-
-	&:hover h1 {
-		outline: 1px solid $light-gray-500;
-		transition: 0.2s outline;
-	}
-
-	&.is-selected h1 {
-		outline: 2px solid $light-gray-500;
-		transition: 0.2s outline;
-	}
-
-	.editor-text-editor & {
-		padding-left: 20px;
-		padding-right: 20px;
-
-		h1 {
-			margin: 0 -10px;
-			padding: 0 10px;
-		}
-
-		@include break-large() {
-			margin: 0 auto;
-			padding: 0;
-		}
-	}
-}
-
-textarea.editor-post-title__input {
-	display: block;
-	font-size: 2em;
-	font-family: $editor-font;
-	line-height: $default-line-height;
-	outline: none;
-	border: none;
-	box-shadow: none;
-	width: 100%;
-	padding: 5px 0;
-	margin: 0;
-
-	&:focus {
-		outline: none;
-		box-shadow: none;
-	}
-
-	.editor-text-editor & {
-		padding: 0;
-	}
-}
